CBL-8389 : Fix build and static analysis issues in kt-serialization API changes (#498)
* Replace var with explicit FLDict type in Document.setContent (common is compiled at Java 8)
* Suppress PMD SingularField and SpotBugs URF_UNREAD_FIELD on Document.extraBackingStore (keep-alive field)
* Add null check on FLValue.fromData() result in Document.setContent (NP_NULL_ON_SOME_PATH)
* Move serialization tests from test/java to test/kotlin so that only the android-ktx test builds compile them :
    - FleeceSerializationTest.kt moved as-is
    - ResultTest.kt renamed to ResultSerializationTest.kt
    - Serialization test and TestModel extracted from CollectionTest.kt into new CollectionSerializationTest.kt
